drop item to token and another token
