+++ dragonbane.js
@@ -83,8 +83,7 @@ async function preloadHandlebarsTemplates() {
         "systems/dragonbane/templates/partials/roll-v12.html",
         "systems/dragonbane/templates/partials/skill-roll-message.hbs",
         "systems/dragonbane/templates/partials/tooltip.hbs",
-        "systems/dragonbane/templates/partials/wp-widget.hbs",
-        "systems/dragonbane/templates/dialogs/pass-item.hbs"
+        "systems/dragonbane/templates/partials/wp-widget.hbs"
     ];
 
     if (game.release.generation < 13) {
@@ -601,70 +600,18 @@ Hooks.on("yzeCombatReady", () => {
 Hooks.on("dropCanvasData", async (canvas,data) => {
 
     if(data.type = "Item" && data.uuid.includes("Actor")){
-        const dropX = data.x;
-        const dropY = data.y;
-        const dropPoint = { x: dropX, y: dropY };
-        const tokens = game.canvas.tokens.placeables;
-        function distance(a, b) {
-            const dx = a.x - b.x;
-            const dy = a.y - b.y;
-            return Math.sqrt(dx * dx + dy * dy);
-        }
-        let closestToken = null;
-        let minDistance = Infinity;
-        for (const token of tokens) {
-            const center = token.center;
-            const dist = distance(center, dropPoint);
-            if (dist < minDistance) {
-                minDistance = dist;
-                closestToken = token;
-            }
-        }
+        await DoD_Utility.dropToAnotherToken(data);
+        return
+    }
+    if(data.type = "Item" && !data.uuid.includes("Actor")){
         const item = await fromUuid(data.uuid);
-        const content = await DoD_Utility.renderTemplate("systems/dragonbane/templates/dialogs/pass-item.hbs",{item: item.name, actor:closestToken.name});
-        const actorID = data.uuid.split(".")[1];
-        const actor = await game.actors.get(actorID);
-        const targetActor = closestToken.actor;
-        const dialog = new Dialog({
-            title: game.i18n.localize("DoD.ui.dialog.passItem"),
-            content: content,
-            buttons:{
-                pass:{
-                    label: game.i18n.localize("DoD.ui.dialog.confirmPassItem"),
-                    callback: async () =>{
-                        if(!game.user.isGM){
-                            if(item.system.quantity > 1){
-                                await item.update({['system.quantity']: item.system.system.quantity - 1})
-                            }
-                              else{
-                                await actor.deleteEmbeddedDocuments("Item", [item])
-                              }
-                            data.target = targetActor;
-                            game.socket.emit("system.dragonbane", {
-                                type: "dropItemtoAnotherCharacter",
-                                data: data
-                            })
-                        }
-                        else{               
-                            if(item.system.quantity > 1){
-                                await item.update({['system.quantity']: item.system.quantity - 1})
-                                const newItem =  game.items.filter(element => element.name === item.name)[0];
-                                await targetActor.createEmbeddedDocuments("Item", [newItem])
-                            }
-                            else{
-                                await actor.deleteEmbeddedDocuments("Item", [data.uuid.split(".")[3]])
-                                await targetActor.createEmbeddedDocuments("Item", [item])
-                            }
+        const itemTypes = ["ability", "armor", "helmet", "injury", "item", "spell", "weapon"];
+        if(itemTypes.includes(item.type)){
+            await DoD_Utility.dropOnToken(data);
+        }
 
-                        }
-                    }
-                },
-                cancel:{label: game.i18n.localize("Cancel")}
-                    }
-                })
-        dialog.render(true)
-            }
-        })
+    }
+})
