+++ modules/socketHandler.mjs
@@ -11,16 +11,16 @@ export default class SocketHandler{
             switch(data.type){
                 case "dropItemtoAnotherCharacter":
                     const item = await fromUuid(data.data.uuid);
-                    const targetActor = data.data.closestToken.actor;
+                    const targetActor = game.actors.get(data.data.target._id);
                     const user = game.user._id;
                     if(targetActor.ownership[user] === 3 && !game.user.isGM){
                         if(item.system.quantity > 1){
                             const newItem = await item.clone();;
                             await newItem.update({['system.quantity']: 1})
-                            await targetActor.createEmbeddedDocuments("Item", newItem)
+                            await targetActor.createEmbeddedDocuments("Item", [newItem])
                         }
                         else{
-                            await targetActor.createEmbeddedDocuments("Item", item)
+                            await targetActor.createEmbeddedDocuments("Item", [item])
                         }
                         return
                     }
@@ -28,10 +28,10 @@ export default class SocketHandler{
                         if(item.system.quantity > 1){
                             const newItem = await item.clone();
                             await newItem.update({['system.quantity']: 1})
-                            await targetActor.createEmbeddedDocuments("Item", newItem)
+                            await targetActor.createEmbeddedDocuments("Item", [newItem])
                         }
                         else{
-                            await targetActor.createEmbeddedDocuments("Item", item)
+                            await targetActor.createEmbeddedDocuments("Item", [item])
                         }
                         return
                     }
