+++ modules/utility.js
@@ -427,5 +427,105 @@ export default class DoD_Utility {
         } else {
             return null;
         }
-    }    
+    } 
+    static async dropToAnotherToken (eventData){
+        const data = eventData;
+        const dropX = data.x;
+        const dropY = data.y;
+        const dropPoint = { x: dropX, y: dropY };
+        const tokens = game.canvas.tokens.placeables;
+        function distance(a, b) {
+            const dx = a.x - b.x;
+            const dy = a.y - b.y;
+            return Math.sqrt(dx * dx + dy * dy);
+        }
+        let closestToken = null;
+        let minDistance = Infinity;
+        for (const token of tokens) {
+            const center = token.center;
+            const dist = distance(center, dropPoint);
+            if (dist < minDistance) {
+                minDistance = dist;
+                closestToken = token;
+            }
+        }
+        const item = await fromUuid(data.uuid);
+
+        const content = game.i18n.format("DoD.ui.dialog.passItemToAnotherCharacter",{item:item.name, character:  closestToken.name})
+        const actorID = data.uuid.split(".")[1];
+        const actor = await game.actors.get(actorID);
+        const targetActor = closestToken.actor;
+        const dialog = new Dialog({
+            title: game.i18n.localize("DoD.ui.dialog.passItem"),
+            content: content,
+            buttons:{
+                pass:{
+                    label: game.i18n.localize("DoD.ui.dialog.confirmPassItem"),
+                    callback: async () =>{
+                        if(!game.user.isGM){
+                           
+                            data.target = targetActor;
+                            game.socket.emit("system.dragonbane", {
+                                type: "dropItemtoAnotherCharacter",
+                                data: data
+                            })
+                             if(item.system.quantity > 1){
+                                await item.update({['system.quantity']: item.system.system.quantity - 1})
+                            }
+                              else{
+                                await actor.deleteEmbeddedDocuments("Item", [data.uuid.split(".")[3]])
+                              }
+                        }
+                        else{               
+                            if(item.system.quantity > 1){
+                                await item.update({['system.quantity']: item.system.quantity - 1})
+                                const newItem =  game.items.filter(element => element.name === item.name)[0];
+                                await targetActor.createEmbeddedDocuments("Item", [newItem])
+                            }
+                            else{
+                                await actor.deleteEmbeddedDocuments("Item", [data.uuid.split(".")[3]])
+                                await targetActor.createEmbeddedDocuments("Item", [item])
+                            }
+
+                        }
+                    }
+                },
+                cancel:{label: game.i18n.localize("Cancel")}
+                    }
+                })
+        dialog.render(true)
+    } 
+    static async dropOnToken(eventData){
+
+        const data = eventData;
+        const dropX = data.x;
+        const dropY = data.y;
+        const dropPoint = { x: dropX, y: dropY };
+        const tokens = game.canvas.tokens.placeables;
+        function distance(a, b) {
+            const dx = a.x - b.x;
+            const dy = a.y - b.y;
+            return Math.sqrt(dx * dx + dy * dy);
+        }
+        let closestToken = null;
+        let minDistance = Infinity;
+        for (const token of tokens) {
+            const center = token.center;
+            const dist = distance(center, dropPoint);
+            if (dist < minDistance) {
+                minDistance = dist;
+                closestToken = token;
+            }
+        }
+        const item = await fromUuid(data.uuid);
+        const targetActor = closestToken.actor;
+        const itemExist = targetActor.items.filter(element => element.name === item.name)[0];
+        if(targetActor.ownership[game.user.id] === 3 || game.user.isGM)
+        if(itemExist){
+            await itemExist.update({['system.quantity']:itemExist.system.quantity + 1 })
+        }
+        else{
+            await targetActor.createEmbeddedDocuments("Item",[item])
+        }
+    }  
 }
